+++ source/DSCResources/AddsOrgUnitsAndGroups/AddsOrgUnitsAndGroups.schema.psm1
@@ -93,8 +93,6 @@ configuration AddsOrgUnitsAndGroups
         Get-OrgUnitSplat -Object $ou -ParentPath $ou.Path -SkipDepend
     }
 
-    $dependencies = @()
-
     foreach ($group in $Groups)
     {
         # remove case sensitivity from hashtables
@@ -102,9 +100,17 @@ configuration AddsOrgUnitsAndGroups
 
         if ($group.GroupScope -eq 'DomainLocal')
         {
-            $dependencies += "[ADGroup]'$($group.GroupName)'"
-            $group.DependsOn = $ouDependencies
             $group.Path = '{0},{1}' -f $group.Path, $DomainDn
+
+            $parentOuRef = "[ADOrganizationalUnit]$($group.Path -replace '\W')"
+            if ($ouDependencies -contains $parentOuRef)
+            {
+                $group.DependsOn = $parentOuRef
+            }
+            else
+            {
+                $group.DependsOn = $ouDependencies
+            }
         }
         elseif (($group.GroupScope -eq 'Global') -or (-not [string]::IsNullOrWhiteSpace($group.Path)))
         {